Extract matching Exceptions to SciJava Ops API
This cleans up OpEnvironment methods
