+++ scijava/scijava-ops-api/src/main/java/org/scijava/ops/api/features/DependencyMatchingException.java
@@ -1,5 +1,5 @@
 
-package org.scijava.ops.engine.matcher;
+package org.scijava.ops.api.features;
 
 import org.scijava.ops.api.OpRef;
