+++ scijava/scijava-ops-engine/src/test/java/org/scijava/ops/engine/hints/AdaptationHintTest.java
@@ -8,9 +8,9 @@ import org.junit.Test;
 import org.scijava.ops.api.Hints;
 import org.scijava.ops.engine.AbstractTestEnvironment;
 import org.scijava.ops.api.OpHints;
+import org.scijava.ops.api.features.OpMatchingException;
 import org.scijava.ops.api.features.BaseOpHints.Adaptation;
 import org.scijava.ops.engine.hint.DefaultHints;
-import org.scijava.ops.engine.matcher.OpMatchingException;
 import org.scijava.ops.spi.OpCollection;
 import org.scijava.ops.spi.OpField;
 import org.scijava.function.Computers;
@@ -38,8 +38,7 @@ public class AdaptationHintTest extends AbstractTestEnvironment {
 			ops.op("test.adaptation.hints").inType(Double[].class).outType(
 				Double[].class).computer();
 			throw new IllegalStateException("This op call should not match!");
-		} catch( IllegalArgumentException e) {
-			assertTrue(e.getCause() instanceof OpMatchingException);
+		} catch( OpMatchingException e) {
 		}
 	}
 
@@ -57,8 +56,7 @@ public class AdaptationHintTest extends AbstractTestEnvironment {
 			ops.op("test.adaptation.hints").inType(Double[].class).outType(
 				Double[].class).computer(hints);
 			throw new IllegalStateException("This op call should not match!");
-		} catch( IllegalArgumentException e) {
-			assertTrue(e.getCause() instanceof OpMatchingException);
+		} catch( OpMatchingException e) {
 		}
 	}
 
@@ -81,8 +79,7 @@ public class AdaptationHintTest extends AbstractTestEnvironment {
 			ops.op("test.adaptation.unadaptable").inType(Double[].class).outType(
 				Double[].class).computer();
 			throw new IllegalStateException("This op call should not match!");
-		} catch( IllegalArgumentException e) {
-			assertTrue(e.getCause() instanceof OpMatchingException);
+		} catch( OpMatchingException e) {
 		}
 	}
 
@@ -100,8 +97,7 @@ public class AdaptationHintTest extends AbstractTestEnvironment {
 			ops.op("test.adaptation.unadaptable").inType(Double[].class).outType(
 				Double[].class).computer(hints);
 			throw new IllegalStateException("This op call should not match!");
-		} catch( IllegalArgumentException e) {
-			assertTrue(e.getCause() instanceof OpMatchingException);
+		} catch( OpMatchingException e) {
 		}
 	}
