+++ scijava/scijava-ops-engine/src/test/java/org/scijava/ops/engine/matcher/DefaultMatchingErrorTest.java
@@ -6,8 +6,9 @@ import java.util.function.Function;
 import org.junit.Assert;
 import org.junit.Test;
 import org.scijava.function.Computers;
+import org.scijava.ops.api.features.DependencyMatchingException;
+import org.scijava.ops.api.features.OpMatchingException;
 import org.scijava.ops.engine.AbstractTestEnvironment;
-import org.scijava.ops.engine.matcher.impl.OpAdaptationInfo;
 import org.scijava.ops.spi.Op;
 import org.scijava.ops.spi.OpCollection;
 import org.scijava.ops.spi.OpDependency;
@@ -34,10 +35,10 @@ public class DefaultMatchingErrorTest extends AbstractTestEnvironment {
 				.function();
 			Assert.fail();
 		}
-		catch (IllegalArgumentException e) {
+		catch (OpMatchingException e) {
 			Assert.assertTrue(e.getMessage().startsWith(
-				"org.scijava.ops.engine.matcher.OpMatchingException: " +
-					"Multiple 'test.duplicateOp/java.util.function.Function<java.lang.Double, java.lang.Double>' " +
+				"Multiple 'test.duplicateOp/" +
+					"java.util.function.Function<java.lang.Double, java.lang.Double>' " +
 					"ops of priority 0.0:"));
 		}
 	}
@@ -52,10 +53,8 @@ public class DefaultMatchingErrorTest extends AbstractTestEnvironment {
 				.apply();
 			Assert.fail("Expected DependencyMatchingException");
 		}
-		catch (IllegalArgumentException e) {
-			Throwable cause = e.getCause();
-			Assert.assertTrue(cause instanceof DependencyMatchingException);
-			String message = cause.getMessage();
+		catch (DependencyMatchingException e) {
+			String message = e.getMessage();
 			Assert.assertTrue(message.contains("Name: \"test.nonexistingOp\""));
 		}
 	}
@@ -70,10 +69,8 @@ public class DefaultMatchingErrorTest extends AbstractTestEnvironment {
 			ops.op("test.outsideOp").input(1.).outType(Double.class).apply();
 			Assert.fail("Expected DependencyMatchingException");
 		}
-		catch (IllegalArgumentException e) {
-			Throwable cause = e.getCause();
-			Assert.assertTrue(cause instanceof DependencyMatchingException);
-			String message = cause.getMessage();
+		catch (DependencyMatchingException e) {
+			String message = e.getMessage();
 			Assert.assertTrue(message.contains("Name: \"test.missingDependencyOp\""));
 			Assert.assertTrue(message.contains("Name: \"test.nonexistingOp\""));
 		}
@@ -90,10 +87,8 @@ public class DefaultMatchingErrorTest extends AbstractTestEnvironment {
 			ops.op("test.adaptMissingDep").input(d).outType(Double[].class).apply();
 			Assert.fail("Expected DependencyMatchingException");
 		}
-		catch (IllegalArgumentException e) {
-			Throwable cause = e.getCause();
-			Assert.assertTrue(cause instanceof DependencyMatchingException);
-			String message = cause.getMessage();
+		catch (DependencyMatchingException e) {
+			String message = e.getMessage();
 			Assert.assertTrue(message.contains("Adaptor:"));
 			Assert.assertTrue(message.contains("Name: \"test.nonexistingOp\""));
